refactor #3084: Remove contacts fields from organization query Removed 'contacts' fields from the 'organization' query in the files 'organization.generated.ts' and 'organization.graphql'. This helps reduce the load on the server and improve application performance by making the request more lightweight and fetching only the most essential organization details. The contacts data is now fetched more efficiently through the new 'organizationPeoplePanel' query.
Add OrganizationPeoplePanel query and simplify PeoplePanel
Created organizationPeoplePanel query to efficiently fetch the essential contact details from the server. As a result, the PeoplePanel and ContactCard components were simplified by eliminating various mutations that would have been called in previous implementation. Additionally, some functionalities and states were moved to the ContactCard component for better encapsulation and code readability. This better aligns with the separation of concerns principle. The changes reduce load on server and improve application performance.
