+++ packages/apps/spaces/app/organizations/[id]/graphql/organization.graphql
@@ -23,35 +23,5 @@ query Organization($id: ID!) {
       relationship
       stage
     }
-    contacts(pagination: { page: 0, limit: 100 }) {
-      content {
-        id
-        name
-        firstName
-        lastName
-        prefix
-        description
-        timezone
-        jobRoles {
-          id
-          primary
-          jobTitle
-          description
-          company
-        }
-        phoneNumbers {
-          id
-          e164
-          rawPhoneNumber
-          label
-          primary
-        }
-        emails {
-          id
-          email
-        }
-      }
-      totalElements
-    }
   }
 }
