+++ packages/apps/spaces/app/organizations/[id]/graphql/organizationPeoplePanel.generated.ts
@@ -1,0 +1,176 @@
+// @ts-nocheck remove this when typscript-react-query plugin is fixed
+import * as Types from '../../../types/__generated__/graphql.types';
+
+import { GraphQLClient } from 'graphql-request';
+import { RequestInit } from 'graphql-request/dist/types.dom';
+import {
+  useQuery,
+  useInfiniteQuery,
+  UseQueryOptions,
+  UseInfiniteQueryOptions,
+} from '@tanstack/react-query';
+
+function fetcher<TData, TVariables extends { [key: string]: any }>(
+  client: GraphQLClient,
+  query: string,
+  variables?: TVariables,
+  requestHeaders?: RequestInit['headers'],
+) {
+  return async (): Promise<TData> =>
+    client.request({
+      document: query,
+      variables,
+      requestHeaders,
+    });
+}
+export type OrganizationPeoplePanelQueryVariables = Types.Exact<{
+  id: Types.Scalars['ID'];
+}>;
+
+export type OrganizationPeoplePanelQuery = {
+  __typename?: 'Query';
+  organization?: {
+    __typename?: 'Organization';
+    id: string;
+    name: string;
+    contacts: {
+      __typename?: 'ContactsPage';
+      totalElements: any;
+      content: Array<{
+        __typename?: 'Contact';
+        id: string;
+        name?: string | null;
+        firstName?: string | null;
+        lastName?: string | null;
+        prefix?: string | null;
+        description?: string | null;
+        timezone?: string | null;
+        jobRoles: Array<{
+          __typename?: 'JobRole';
+          id: string;
+          primary: boolean;
+          jobTitle?: string | null;
+          description?: string | null;
+          company?: string | null;
+        }>;
+        phoneNumbers: Array<{
+          __typename?: 'PhoneNumber';
+          id: string;
+          e164?: string | null;
+          rawPhoneNumber?: string | null;
+          label?: Types.PhoneNumberLabel | null;
+          primary: boolean;
+        }>;
+        emails: Array<{
+          __typename?: 'Email';
+          id: string;
+          email?: string | null;
+        }>;
+      }>;
+    };
+  } | null;
+};
+
+export const OrganizationPeoplePanelDocument = `
+    query OrganizationPeoplePanel($id: ID!) {
+  organization(id: $id) {
+    id
+    name
+    contacts(pagination: {page: 0, limit: 100}) {
+      content {
+        id
+        name
+        firstName
+        lastName
+        prefix
+        description
+        timezone
+        jobRoles {
+          id
+          primary
+          jobTitle
+          description
+          company
+        }
+        phoneNumbers {
+          id
+          e164
+          rawPhoneNumber
+          label
+          primary
+        }
+        emails {
+          id
+          email
+        }
+      }
+      totalElements
+    }
+  }
+}
+    `;
+export const useOrganizationPeoplePanelQuery = <
+  TData = OrganizationPeoplePanelQuery,
+  TError = unknown,
+>(
+  client: GraphQLClient,
+  variables: OrganizationPeoplePanelQueryVariables,
+  options?: UseQueryOptions<OrganizationPeoplePanelQuery, TError, TData>,
+  headers?: RequestInit['headers'],
+) =>
+  useQuery<OrganizationPeoplePanelQuery, TError, TData>(
+    ['OrganizationPeoplePanel', variables],
+    fetcher<
+      OrganizationPeoplePanelQuery,
+      OrganizationPeoplePanelQueryVariables
+    >(client, OrganizationPeoplePanelDocument, variables, headers),
+    options,
+  );
+useOrganizationPeoplePanelQuery.document = OrganizationPeoplePanelDocument;
+
+useOrganizationPeoplePanelQuery.getKey = (
+  variables: OrganizationPeoplePanelQueryVariables,
+) => ['OrganizationPeoplePanel', variables];
+export const useInfiniteOrganizationPeoplePanelQuery = <
+  TData = OrganizationPeoplePanelQuery,
+  TError = unknown,
+>(
+  pageParamKey: keyof OrganizationPeoplePanelQueryVariables,
+  client: GraphQLClient,
+  variables: OrganizationPeoplePanelQueryVariables,
+  options?: UseInfiniteQueryOptions<
+    OrganizationPeoplePanelQuery,
+    TError,
+    TData
+  >,
+  headers?: RequestInit['headers'],
+) =>
+  useInfiniteQuery<OrganizationPeoplePanelQuery, TError, TData>(
+    ['OrganizationPeoplePanel.infinite', variables],
+    (metaData) =>
+      fetcher<
+        OrganizationPeoplePanelQuery,
+        OrganizationPeoplePanelQueryVariables
+      >(
+        client,
+        OrganizationPeoplePanelDocument,
+        { ...variables, ...(metaData.pageParam ?? {}) },
+        headers,
+      )(),
+    options,
+  );
+
+useInfiniteOrganizationPeoplePanelQuery.getKey = (
+  variables: OrganizationPeoplePanelQueryVariables,
+) => ['OrganizationPeoplePanel.infinite', variables];
+useOrganizationPeoplePanelQuery.fetcher = (
+  client: GraphQLClient,
+  variables: OrganizationPeoplePanelQueryVariables,
+  headers?: RequestInit['headers'],
+) =>
+  fetcher<OrganizationPeoplePanelQuery, OrganizationPeoplePanelQueryVariables>(
+    client,
+    OrganizationPeoplePanelDocument,
+    variables,
+    headers,
+  );
